Give highest precedence to function calls
Previously, expression like "exp(1)^2" would get parsed as "exp((1)^2)".
This is logically consistent if we treat `exp` as a unary operator with
precedence lower than the `^` binary operator, but it's also very
counter-intuitive - the user probably meant to say "(exp(1))^2".
To solve this issue, we will treat parenthesised function calls as
having precedence higher than binary operators. This causes the `exp`
example to parse in the expected way.
To avoid confusion, the unary `-` followed by left parenthesis will
*not* be considered a function call. "-(-1)^2" still parses as
"- ((-1)^2)" == "-1", which (probably) is a more reasonable behaviour.